solve day four part one

+++ src/aoc/days/eight.rs
@@ -3,7 +3,7 @@ use crate::aoc::Solution;
 pub struct Eight;
 
 impl Solution for Eight {
-    type Output = usize;
+    type Output = u32;
     type Parsed = Vec<u32>;
 
     fn input() -> &'static str {
