+++ src/aoc/days/twenty_one.rs
@@ -3,7 +3,7 @@ use crate::aoc::Solution;
 pub struct TwentyOne;
 
 impl Solution for TwentyOne {
-    type Output = usize;
+    type Output = u32;
     type Parsed = Vec<u32>;
 
     fn input() -> &'static str {
